packages/pocketbase_backend/src/pocketbase/utils.py
"""
Utility functions for PocketBase API.
"""

from dotenv import load_dotenv  # type: ignore


def load_env() -> None:
    """
    Loads environment variables from a .env file.
    """
    load_dotenv()
