packages/pocketbase_backend/pyproject.toml
[tool.setuptools.packages.find]
where = ["src"]
[project]
name = "pocketbase"
version = "0.1.0"
description = "PocketBase binary manager for Python projects."
requires-python = ">=3.8"
dependencies = [
    "python-dotenv"
]

[project.scripts]
pocketbase-manager = "pocketbase.pocketbase_manager:main"

